Add more comments for refactoring
Moving further down the timing loop I add more refactoring comments as
hooks. I will begin by refactoring into Controllers and Stations objects
and will pick up the Program object as needed. It appears that the
notion of a board in this code is what I have declared to be Controller
which was being determined purely by the range of station ID's.