Refactor program duration update logic for stations
The existing implementation had a rather indirect way of updating the
run schedule and program schedule gv structures. This resulted in a
duplication of code logic. I refactored the conditional to include the
check for concurrency inline with duration confirmation for continuing
to before the runschedule and programschedule updates. The resulting
output is functionally identical but easier to reason about.