Add master station calls to gv API
Added functions to assist in dealing with master stations to the gv API.
IndexOfMaster() : Int returns the index of the master station
designated. MasterStationAssigned() : bool returns if the index of the
master station is non-zero indicating that a station has master
designation. StationIsMaster(station_id) : bool informs the caller if
the provided station index is a master station. Refactored sip.py to use
new helper methods and added to reference.txt document.