Add gv API calls to clear UI program data
Allow the caller to clear the UI program schedule for a station with
ClearUIProgramScheduleForStation(station_index) or for all stations with
ClearUIProgramScheduleForAllStations(). This allows for a single point
of change to these algorithms. Refactor sip.py and helpers.py to use the
new calls and add documentation to reference file.